# YJIT: Pass method arguments using registers