--- a/yjit/src/backend/ir.rs
+++ b/yjit/src/backend/ir.rs
@@ -1266,11 +1266,11 @@ impl Assembler
 
     /// Spill all live registers to the stack
     pub fn spill_regs(&mut self) {
-        self.spill_regs_except(vec![]);
+        self.spill_regs_except(&vec![]);
     }
 
     /// Spill all live registers except `ignored_temps` to the stack
-    pub fn spill_regs_except(&mut self, ignored_temps: Vec<RegOpnd>) {
+    pub fn spill_regs_except(&mut self, ignored_temps: &Vec<RegOpnd>) {
         // Forget registers above the stack top
         let mut reg_mapping = self.ctx.get_reg_mapping();
         for stack_idx in self.ctx.get_stack_size()..MAX_CTX_TEMPS as u8 {
@@ -1346,6 +1346,42 @@ impl Assembler
         }
     }
 
+    // Shuffle register moves, sometimes adding extra moves using SCRATCH_REG,
+    // so that they will not rewrite each other before they are used.
+    pub fn reorder_reg_moves(old_moves: &Vec<(Reg, Opnd)>) -> Vec<(Reg, Opnd)> {
+        // Return the index of a move whose destination is not used as a source if any.
+        fn find_safe_move(moves: &Vec<(Reg, Opnd)>) -> Option<usize> {
+            moves.iter().enumerate().find(|(_, &(dest_reg, _))| {
+                moves.iter().all(|&(_, src_opnd)| src_opnd != Opnd::Reg(dest_reg))
+            }).map(|(index, _)| index)
+        }
+
+        // Remove moves whose source and destination are the same
+        let mut old_moves: Vec<(Reg, Opnd)> = old_moves.clone().into_iter()
+            .filter(|&(reg, opnd)| Opnd::Reg(reg) != opnd).collect();
+
+        let mut new_moves = vec![];
+        while old_moves.len() > 0 {
+            // Keep taking safe moves
+            while let Some(index) = find_safe_move(&old_moves) {
+                new_moves.push(old_moves.remove(index));
+            }
+
+            // No safe move. Load the source of one move into SCRATCH_REG, and
+            // then load SCRATCH_REG into the destination when it's safe.
+            if old_moves.len() > 0 {
+                // Make sure it's safe to use SCRATCH_REG
+                assert!(old_moves.iter().all(|&(_, opnd)| opnd != Opnd::Reg(Assembler::SCRATCH_REG)));
+
+                // Move SCRATCH <- opnd, and delay reg <- SCRATCH
+                let (reg, opnd) = old_moves.remove(0);
+                new_moves.push((Assembler::SCRATCH_REG, opnd));
+                old_moves.push((reg, Opnd::Reg(Assembler::SCRATCH_REG)));
+            }
+        }
+        new_moves
+    }
+
     /// Sets the out field on the various instructions that require allocated
     /// registers because their output is used as the operand on a subsequent
     /// instruction. This is our implementation of the linear scan algorithm.
@@ -1389,42 +1425,6 @@ impl Assembler
             if let Some(reg_index) = reg_index {
                 *pool &= !(1 << reg_index);
             }
-        }
-
-        // Reorder C argument moves, sometimes adding extra moves using SCRATCH_REG,
-        // so that they will not rewrite each other before they are used.
-        fn reorder_c_args(c_args: &Vec<(Reg, Opnd)>) -> Vec<(Reg, Opnd)> {
-            // Return the index of a move whose destination is not used as a source if any.
-            fn find_safe_arg(c_args: &Vec<(Reg, Opnd)>) -> Option<usize> {
-                c_args.iter().enumerate().find(|(_, &(dest_reg, _))| {
-                    c_args.iter().all(|&(_, src_opnd)| src_opnd != Opnd::Reg(dest_reg))
-                }).map(|(index, _)| index)
-            }
-
-            // Remove moves whose source and destination are the same
-            let mut c_args: Vec<(Reg, Opnd)> = c_args.clone().into_iter()
-                .filter(|&(reg, opnd)| Opnd::Reg(reg) != opnd).collect();
-
-            let mut moves = vec![];
-            while c_args.len() > 0 {
-                // Keep taking safe moves
-                while let Some(index) = find_safe_arg(&c_args) {
-                    moves.push(c_args.remove(index));
-                }
-
-                // No safe move. Load the source of one move into SCRATCH_REG, and
-                // then load SCRATCH_REG into the destination when it's safe.
-                if c_args.len() > 0 {
-                    // Make sure it's safe to use SCRATCH_REG
-                    assert!(c_args.iter().all(|&(_, opnd)| opnd != Opnd::Reg(Assembler::SCRATCH_REG)));
-
-                    // Move SCRATCH <- opnd, and delay reg <- SCRATCH
-                    let (reg, opnd) = c_args.remove(0);
-                    moves.push((Assembler::SCRATCH_REG, opnd));
-                    c_args.push((reg, Opnd::Reg(Assembler::SCRATCH_REG)));
-                }
-            }
-            moves
         }
 
         // Adjust the number of entries in live_ranges so that it can be indexed by mapped indexes.
@@ -1602,7 +1602,7 @@ impl Assembler
                 if c_args.len() > 0 {
                     // Resolve C argument dependencies
                     let c_args_len = c_args.len() as isize;
-                    let moves = reorder_c_args(&c_args.drain(..).into_iter().collect());
+                    let moves = Self::reorder_reg_moves(&c_args.drain(..).into_iter().collect());
                     shift_live_ranges(&mut shifted_live_ranges, asm.insns.len(), moves.len() as isize - c_args_len);
 
                     // Push batched C arguments