@maximecb 2024-08-09
Hi Kokubun.

Congratulations on getting all the tests passing on this. I'm sure it was not easy. At first glance I think the PR makes sense. I think I understand most of it, but I did ask some clarifying questions.

My main concern would be code size, if we are generating more versions to match register allocation. Can you check the code size before and after on lobsters? I'm also curious to see how performance changes on `fib` and `30k_methods` (because those benchmarks exercise calls heavily and may amplify the difference).

I wish we had gotten more of a speedup from this change. It would seem like being able to access arguments through registers should be a decent win. It may be worth looking at the machine code to see how efficient things are. Not opposed to merging if we think that code size is not an issue and if we think we have a path forward for improving code generation more.