@k0kubun 2024-08-12
> My main concern would be code size, if we are generating more versions to match register allocation. Can you check the code size before and after on lobsters?

Copying from https://github.com/ruby/ruby/pull/11280#discussion_r1714115804, 

Before (master):

```
compiled_block_count:         70,803
versions_per_block:            1.501
inline_code_size:          7,503,452
outlined_code_size:        6,436,718
code_region_size:         15,413,248
```

After (this PR):

```
compiled_block_count:         75,454
versions_per_block:            1.600
inline_code_size:          8,359,786
outlined_code_size:        6,892,488
code_region_size:         17,068,032
```

> I'm also curious to see how performance changes on fib and 30k_methods (because those benchmarks exercise calls heavily and may amplify the difference).

It seems to generate code for `fib` as I intended, but the speedup is just 2%. With so many instructions for a frame push and removing or optimizing only a few instructions in the block, I'm not that surprised by the result though.

On the contrary, `30k_methods` seems to have a weird slowdown. I'll look into it. 

```
before: ruby 3.4.0dev (2024-08-08T23:13:16Z master 77ffdfe79f) +YJIT [x86_64-linux]
after: ruby 3.4.0dev (2024-08-09T03:18:25Z yjit-local-args a459c77cae) +YJIT [x86_64-linux]

-----------  -----------  ----------  ----------  ----------  -------------  ------------
bench        before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
30k_methods  352.1        1.2         363.7       0.1         0.98           0.97
fib          20.2         0.7         19.8        0.3         0.99           1.02
-----------  -----------  ----------  ----------  ----------  -------------  ------------
```

> I wish we had gotten more of a speedup from this change. It would seem like being able to access arguments through registers should be a decent win. It may be worth looking at the machine code to see how efficient things are.

I also feel like there might be some unnoticed trick that would realize its full potential across the board, but so far, nothing in generated code seems wrong.

It seems like I confused you in the weekly meeting that I discovered a bug that I haven't fixed yet, but this PR already has the fix for it. The `&& version_ctx.reg_mapping == ctx.reg_mapping` trick https://github.com/ruby/ruby/pull/11280#discussion_r1714115804 indeed sped up `activerecord` by a lot, compared to before fixing it, but it unfortunately didn't have much impact on other benchmarks.

I'll try to keep improving the implementation, but I have a suspicion that we have too many C calls/spills to benefit from this, or that the current bottleneck is not here and thus it seems insignificant as of today.

Nevertheless, I still want to move forward with this patch merged so that we can stop thinking about "our JIT code is slow because we haven't even optimized locals yet".

---

Now I thought we can use our time more efficiently if we pair over this first, so I sent an invite.