@maximecb 2024-08-13
> Nevertheless, I still want to move forward with this patch merged so that we can stop thinking about "our JIT code is slow because we haven't even optimized locals yet".

To me the code size is the main worry right now. It's about an 11% increase in inline code size. This is big. Might be worth looking at whether instruction cache misses increase.

It would also make sense to check if there are too many spills as you suggest. Ideally code using more registers should be smaller (less loads/stores to/from memory), so the question is, does the increase just come from more versions, or is it because there are many more loads and spills.

> I also feel like there might be some unnoticed trick that would realize its full potential across the board, but so far, nothing in generated code seems wrong.

I think it's likely that with some targeted tweaks you might be able to tip the balance such that it speeds up most benchmarks.