@maximecb 2024-08-14
Something occurred to me yesterday.

I think it would be worth trying to get more insight into where extra blocks gets created.  We have a 6.5% increased in compiled block count. What worries me is that this increase could be very unevenly distributed. What I mean is that most methods may not see any increase in block version count, but you could get into a situation where some loop or some method gets duplicated many times over, which would be potentially very bad in terms of instruction cache performance. It's worth digging a bit to make sure that is not the case.

I apologize for making your life difficult Kokubun 😅 . Code size is really my only concern here, the only obstacle to merging. So I think it would make sense to try to look for pathological cases in our benchmarks and try to convince ourselves that there are none, or find ways to mitigate them if there are. Might be worth looking at protoboeuf since it has very big generated methods.

Back when I implemented BBV for my masters, I had set it up so that register allocation was not accounted for when looking for a compatible version. The downside is that this requires generating moves to go from versions with incompatible register allocation. The upside is that there is no increase in the number of versions (though I can't speak about the effect on code size). Alternatively, it may be possible to use heuristics to try to make the register allocation at the input of blocks match up better so that less versions are created.