@k0kubun 2024-08-14
> generating moves to go from versions with incompatible register allocation

Pairing with Alan and Randy yesterday, Alan suggested the same thing. I've just finished what we left off yesterday, and pushed a commit to generate such moves https://github.com/ruby/ruby/pull/11280/commits/974f465cab5fc86e6533c8a2a5058d802d64aa60.

It decreased the number of versions generated, and it didn't slow down. Here's the code size comparison on lobsters (Note: I used `YJIT.runtime_stats` in the harness with `--yjit`, instead of using `--yjit-stats` which counts stats code too):

master:

```
compiled_block_count: 68127
inline_code_size: 4575579
outlined_code_size: 3957747
code_region_size: 11010048
```

Before (this PR without register moves):

```
compiled_block_count: 72293
inline_code_size: 4817514
outlined_code_size: 4510070
code_region_size: 11759616
```

After (this PR with register moves):

```
compiled_block_count: 69805
inline_code_size: 4669535
outlined_code_size: 4232636
code_region_size: 11366400
```

Compared to master, `inline_code_size` is increased only by 2%, which was 5.3% without register moves. `inline_code_size / compiled_block_count` is down from 67.16 to 66.89, so I think we're generating more compact code for each block.