@k0kubun 2024-08-15
> Can you post the performance difference on headline as well as 30k_methods, fib and protoboeuf?

Here's the latest result:

```
before: ruby 3.4.0dev (2024-08-15T20:00:09Z master abc04e898b) +YJIT [x86_64-linux]
after: ruby 3.4.0dev (2024-08-15T20:19:50Z yjit-local-args 7bd011cc8b) +YJIT [x86_64-linux]

--------------  -----------  ----------  ----------  ----------  -------------  ------------
bench           before (ms)  stddev (%)  after (ms)  stddev (%)  after 1st itr  before/after
activerecord    117.3        0.4         118.6       0.2         0.98           0.99
chunky-png      292.3        0.3         293.4       0.3         1.00           1.00
erubi-rails     509.7        0.4         502.9       0.4         0.99           1.01
hexapdf         994.6        1.0         997.3       1.1         1.00           1.00
liquid-c        31.6         0.8         31.1        0.7         0.97           1.02
liquid-compile  34.1         0.6         34.2        1.0         0.93           1.00
liquid-render   43.3         0.5         42.6        0.6         0.96           1.02
lobsters        504.9        1.0         508.3       2.2         0.97           0.99
mail            69.6         0.3         68.3        0.5         1.01           1.02
psych-load      954.3        0.1         958.1       0.1         0.99           1.00
railsbench      1207.3       0.9         1210.5      1.3         0.99           1.00
rubocop         62.1         5.6         62.3        5.7         0.98           1.00
ruby-lsp        68.9         0.8         72.2        1.2         0.97           0.96
sequel          33.9         0.5         33.6        0.3         1.01           1.01
protoboeuf      26.2         0.2         26.2        0.1         0.97           1.00
30k_methods     352.8        0.2         360.6       0.2         0.98           0.98
fib             20.2         0.6         19.8        0.2         1.02           1.02
--------------  -----------  ----------  ----------  ----------  -------------  ------------
```

Three benchmarks are sped up by 2%, so I think the result is still positive. However, `ruby-lsp` slowed down a lot this time, so I'll look into the generated code.